+++ lib/system.rb
@@ -6,9 +6,9 @@ module System
 
     if (!fire_forget)
       if system cmd
-        puts "-> Action [#{action_message}] complete\n\n"
+        puts "\nAction #{action_message} complete".black.on_light_green
       else
-        raise "\n!!!\n   Error trying to #{action_message}\n!!!\n\n"
+        raise "\n!!!\n   Error trying to #{action_message}\n!!!\n\n".red
       end
     else
       system cmd